ml/training/recommendation-model/requirements.txt
lightfm==1.17
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4
mlflow==2.8.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9

